app/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
azure-cognitiveservices-speech
pandas
openpyxl
websockets
python-dotenv
ffmpeg
